requirements.txt
openai 
httpx==0.27.2
pandas
numpy
sentence-transformers
faiss-cpu
python-dotenv